Переделал и исправил создание и вывод пользователя
