+++ WebApplicationShopOnline/Controllers/UserController.cs
@@ -5,10 +5,10 @@ namespace WebApplicationShopOnline.Controllers
 {
 	public class UserController : Controller
 	{
-		public IActionResult Index()
+		public IActionResult Index(string name, string login, string password, string phone, string email)
 		{
-			UserViewModel testUser = new UserViewModel("Test", "keks@mail.com", "password", "+77777777");
-			return View(testUser);
+			User inputUser = new User(name, login, password, phone, email);
+			return View(inputUser);
 		}
 	}
 }
