+++ WebApplicationShopOnline/Views/Home/Index.cshtml
@@ -13,6 +13,7 @@
                 <li>Создание контроллера</li>
                 <li>Создание представления</li>
                 <li><a asp-controller="User" asp-action="Index"> Создание класса пользователь с полями: имя, логин, пароль, телефон. Создание полного MVC  для отображения информации о пользователе</a></li>
+                <li>Скопируйте параметры для теста: name=stas&login=stas101&password=Aa123456&phone=+7928777777&email=eele@main.com</li>
             </ul>
         </li>
         <li>4.02.2025
