+++ WebApplicationShopOnline/Views/Product/Index.cshtml
@@ -1,0 +1,10 @@
+﻿@model Product
+
+<div class="card">
+	<div class="card-body">
+		@Model.Name
+	</div>
+	<div class="card-footer">
+		@Model.Cost
+	</div>
+</div>
